src/requirements.txt
fastapi[standard]
groq
langfuse
langchain
langchain-groq==0.3.1
langchain-openai>=0.2.0
langgraph>=0.2.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
pydantic[email]>=2.3.0
bcrypt==4.0.1
redis>=4.2.0,<5.0.0
aioredis==2.0.1
fastapi-limiter==0.1.5
pymysql==1.1.1 
google-genai==1.16.1
google-generativeai==0.8.5
python-dotenv>=1.0.0
httpx>=0.25.0
PyJWT>=2.8.0

# Audio Monitor Service Dependencies
cryptography>=41.0.0
motor>=3.3.0
assemblyai>=0.17.0
openai>=1.3.0
google-cloud-storage>=2.10.0
paramiko>=3.3.0